내부 설계 재설정

+++ src/main/java/gallery/back/art/backend/api/account/repository/AccountRepository.java
@@ -3,7 +3,7 @@ package gallery.back.art.backend.api.account.repository;
 import gallery.back.art.backend.api.account.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountRepository extends JpaRepository<Member, Integer> {
+public interface AccountRepository extends JpaRepository<Member, Long> {
     Member findByEmailAndPassword(String email, String password);
     Member findByEmail(String email);
 }
